# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Handwork Marketplace API",
    "description": "API for the Handwork Marketplace platform connecting clients with skilled workers",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "dev-support@handworkmarketplace.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.handworkmarketplace.com",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.handworkmarketplace.com",
      "description": "Staging server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Kitchen Sink Repair"
          },
          "description": {
            "type": "string",
            "example": "Need to fix a leaking kitchen sink"
          },
          "category": {
            "type": "string",
            "example": "plumbing"
          },
          "budget_min": {
            "type": "number",
            "format": "float",
            "example": 100.00
          },
          "budget_max": {
            "type": "number",
            "format": "float",
            "example": 200.00
          },
          "location": {
            "type": "string",
            "example": "123 Main St, New York, NY"
          },
          "preferred_date": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["open", "assigned", "in_progress", "completed", "cancelled"],
            "example": "open"
          },
          "client_id": {
            "type": "integer",
            "example": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "status_code": {
            "type": "integer",
            "example": 400
          },
          "details": {
            "type": "object"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get list of jobs",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "budget_min",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "budget_max",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "pages": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Jobs"],
        "summary": "Create a new job",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "category", "budget_min", "budget_max", "location"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "budget_min": {
                    "type": "number",
                    "format": "float"
                  },
                  "budget_max": {
                    "type": "number",
                    "format": "float"
                  },
                  "location": {
                    "type": "string"
                  },
                  "preferred_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "requirements": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get job details",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}